% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Calculate distances between variables.}
\usage{
calculateDistance(x, degrees = 2)
}
\arguments{
\item{x}{A data frame of numeric variables.}

\item{degrees}{The degree to which the differences are raised (e.g., Euclidean distance corresponds to degrees = 2).}
}
\value{
A data frame where each element represents the distance raised to the specified degree.
}
\description{
This function computes the distance between variables in a data frame by raising the differences to the specified power (degrees).
}
\examples{
\dontrun{
  data <- data.frame(var1 = rnorm(10), var2 = rnorm(10))
  distance_matrix <- calculateDistance(data, degrees = 2)
  print(distance_matrix)
}
}
