% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDP2.R
\name{calculateDP2}
\alias{calculateDP2}
\title{Calculate the DP2 composite indicator.}
\usage{
calculateDP2(x, xFactores, iteration = 1)
}
\arguments{
\item{x}{A data frame containing the normalized variables.}

\item{xFactores}{A vector of weights for each variable.}

\item{iteration}{The current iteration number (for labeling purposes).}
}
\value{
A list containing the composite indicator and the weighted matrix.
}
\description{
This function computes the composite indicator based on the weighted factors obtained from MARS.
}
\examples{
\dontrun{
  normalized_data <- data.frame(var1 = rnorm(10), var2 = rnorm(10))
  weights <- c(0.5, 0.5)
  dp2_result <- calculateDP2(normalized_data, weights)
  print(dp2_result[[1]])
}
}
