% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFrechet.R
\name{indexFrechet}
\alias{indexFrechet}
\title{Compute the Fréchet index for composite indicators.}
\usage{
indexFrechet(x, degrees = 2)
}
\arguments{
\item{x}{A data frame of distances calculated from the input variables.}

\item{degrees}{The degree to which the distances are raised (typically used in calculateDistance).}
}
\value{
A vector representing the Fréchet index for each observation.
}
\description{
This function calculates the Fréchet index, which aggregates the distances of the variables.
}
\examples{
\dontrun{
  distance_matrix <- matrix(rnorm(50), nrow = 10)
  frechet_index <- indexFrechet(distance_matrix, degrees = 2)
  print(frechet_index)
}
}
