% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalize the data based on variable polarity.}
\usage{
normalization(x, polarity = NULL, custom_max = NULL, custom_min = NULL)
}
\arguments{
\item{x}{A data frame containing the variables to be normalized.}

\item{polarity}{A vector specifying the columns with positive polarity. All other columns are considered to have negative polarity}

\item{custom_max}{Optional vector of maximum values for each column. If NULL, the maximum is calculated automatically.}

\item{custom_min}{Optional vector of minimum values for each column. If NULL, the minimum is calculated automatically.}
}
\value{
A normalized data frame where each variable is scaled between 0 and 1.
}
\description{
This function normalizes a data frame according to the specified polarity of the variables.
Positive polarity variables are normalized using the Min-Max method, while negative polarity
variables are normalized using the Max-Min method.
}
\examples{
\dontrun{
  data <- data.frame(var1 = rnorm(10), var2 = rnorm(10))
  normalized_data <- normalization(data, polarity = c(1), custom_max = NULL, custom_min = NULL)
  print(normalized_data)
}
}
